src/flask_dash/requirements.txt
anyio==4.10.0
attrs==25.4.0
blinker==1.9.0
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.3
cycler==0.12.1
fiona==1.10.1
Flask==2.3.2
folium==0.14.0
fonttools==4.60.1
geopandas==1.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.9.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.3.1
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==2.3.6
mysql-connector==2.2.9
mysql-connector-python==9.5.0
